+++ src/main/java/union/commands/strikes/DeleteStikeCmd.java
@@ -1,4 +1,4 @@
-package union.commands.moderation;
+package union.commands.strikes;
 
 import java.time.Instant;
 import java.time.temporal.ChronoUnit;
